+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix support of maps with a region crossing the antimeridian (#114)
+
 ## [0.2.0] - 2026-05-19
 
 ### Added